refactor: move data_types.py file to google/cloud/spanner_v1/
