+++ google/cloud/spanner_v1/types/__init__.py
@@ -62,7 +62,6 @@ from .type import (
     Type,
     TypeCode,
 )
-from .data_types import JsonObject
 
 __all__ = (
     "CommitResponse",
@@ -102,5 +101,4 @@ __all__ = (
     "StructType",
     "Type",
     "TypeCode",
-    "JsonObject",
 )
